+++ varset_deprecated.go
@@ -50,7 +50,9 @@ func deprecated[T any, Schema variable.TypedSchema[T]](
 	return deprecatedFunc[T]{
 		[]variable.Any{v},
 		func() (T, bool, error) {
-			return v.NativeValue()
+			return v.NativeValue(),
+				v.Availability() == variable.AvailabilityOK,
+				v.Error()
 		},
 	}
 }